feat: :white_check_mark: create test page to verify local DB storage and use UUIDs for IDs
We add a test page  to verify that messages are correctly stored in the local database with dexie.js. The test page includes functionality to handle message input, recieved messages we didnt use the Componenent because he stored every input in our localDB because we wrote our recieved messages by ourselfs. UUIDs will be used for message IDs.